code commit

+++ 1_C++/1_Basic/4_IMP Code/1_LogicWaleCode/6_Factorial.cpp
@@ -3,7 +3,7 @@ using namespace std;
 
 int main() {
     int n;
-    long long fact = 1;
+    float fact = 1;
     cin >> n; // User input
     for(int i = 1; i <= n; i++) {
         fact *= i; // Multiply current number
